Moves fake services to dedicated folder
