+++ Neolution.Extensions.DataSeeding.UnitTests/AbstractSeedTests.cs
@@ -6,6 +6,7 @@
     using Neolution.Extensions.DataSeeding.Abstractions;
     using Neolution.Extensions.DataSeeding.UnitTests.Fakes;
     using Neolution.Extensions.DataSeeding.UnitTests.Fakes.MultiTenantSeeds;
+    using Neolution.Extensions.DataSeeding.UnitTests.Fakes.Services;
     using Shouldly;
     using Xunit;
     using Xunit.Abstractions;
