+++ Neolution.Extensions.DataSeeding.UnitTests/Fakes/Services/IFakeScopedService.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.Extensions.DataSeeding.UnitTests.Fakes.Services
+{
+    /// <summary>
+    /// Fake scoped service interface.
+    /// </summary>
+    public interface IFakeScopedService
+    {
+        /// <summary>
+        /// Gets the service identifier.
+        /// </summary>
+        string ServiceId { get; }
+    }
+}
