+++ Neolution.Extensions.DataSeeding.UnitTests/Fakes/Services/IFakeSingletonService.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.Extensions.DataSeeding.UnitTests.Fakes.Services
+{
+    /// <summary>
+    /// Fake singleton service interface.
+    /// </summary>
+    public interface IFakeSingletonService
+    {
+        /// <summary>
+        /// Gets the service identifier.
+        /// </summary>
+        string ServiceId { get; }
+    }
+}
