+++ Neolution.Extensions.DataSeeding.UnitTests/Fakes/Services/IFakeTransientService.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.Extensions.DataSeeding.UnitTests.Fakes.Services
+{
+    /// <summary>
+    /// Fake transient service interface.
+    /// </summary>
+    public interface IFakeTransientService
+    {
+        /// <summary>
+        /// Gets the service identifier.
+        /// </summary>
+        string ServiceId { get; }
+    }
+}
